changed busyWait to use clock timers

+++ src/timers.h
@@ -8,5 +8,6 @@
 
 void initTimers(void);
 
+
 /* Delay for a period of time in microseconds */
-void busyWait(uint32 delay);
+void busyWait(uint16 delay);
